requirements.txt
Flask
supabase
python-dotenv
requests
Pillow
Flask-APScheduler
gunicorn